pyproject.toml
[tool.poetry]
name = "pdf2dxf"
version = "0.1.0"
description = "Convert from PDF to DXF"
authors = ["Matteo Bertini <naufraghi@develer.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.7"
dxfwrite = "^1.2"
lxml = "^4.3"
python-svg-parser = {git = "git@github.com:naufraghi/svg.git"}

[tool.poetry.dev-dependencies]
yapf = "^0.26.0"
isort = "^4.3"
flake8 = "^3.7"
mypy-mypyc = "^0.670.0"
pytest = "^4.3"
pytest-mypy = "^0.3.2"
pyinstaller = "^3.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
